+++ spec/apir/examples/default_headers_override_spec.rb
@@ -39,7 +39,14 @@ describe 'Headers override' do
   # when headers are already being constructed
   it 'prepare cookies override' do
     url = "https://prepare_cookies_override.com"
-    stub_request(:any, url).to_return { |request| { body: json_response, headers: request.headers } }
+
+    stub_request(:post, url).
+      with(headers: { 'Cookie': 'auth=session; auth_cookie=some_uid' }).
+      to_return(:status => 200, :body => "success", :headers => { 'Set-Cookie' => 'auth=session' })
+
+    stub_request(:get, url).
+      with(:headers => { 'Cookie' => 'auth_cookie=some_uid' }).
+      to_return(:status => 200, :body => "", :headers => { 'Set-Cookie' => 'auth=session' })
 
     class RedifinedCookieRequest < Apir::Request
       def some_condition
@@ -48,24 +55,43 @@ describe 'Headers override' do
 
       def some_cookie
         HTTP::Cookie.new(
-          name:   'cookie_name',
-          value:  'cookie_value',
-          domain: '.prepare_cookies_override.com',
+          name:   'auth_cookie',
+          value:  'some_uid',
+          domain: 'prepare_cookies_override.com',
           path:   '/'
         )
       end
 
-      def prepare_cookies
-        if some_condition
-          @cookie_jar.add(some_cookie)
-        end
-        @cookie_jar
-      end
+      #   def prepare_cookies
+      #     puts 'prepating cookies overriden'
+      #     if some_condition
+      #       # @cookie_jar.add(some_cookie)
+      #     end
+      #     puts @cookie_jar.cookies
+      #     @cookie_jar
+      #   end
     end
 
-    request = RedifinedCookieRequest.new(url)
+    request        = RedifinedCookieRequest.new(url)
+    second_request = RedifinedCookieRequest.new(url)
 
-    request.post!
-    expect(request.raw_response.cookies).to include('cookie_name' => 'cookie_value')
+    request.cookie_jar << HTTP::Cookie.new(
+      name:   'auth_cookie',
+      value:  'some_uid',
+      domain: '.prepare_cookies_override.com',
+      path:   '/')
+
+    expect(request.cookies).to eq({ 'auth_cookie' => 'some_uid' })
+    second_request.cookie_jar = request.cookie_jar
+    expect(second_request.cookies).to eq({ 'auth_cookie' => 'some_uid' })
+
+    request.get!
+    expect(request.raw_response.cookies).to include('auth' => 'session')
+
+    second_request.cookie_jar = request.raw_response.cookie_jar
+    expect(second_request.cookies).to eq({ 'auth_cookie' => 'some_uid', 'auth' => 'session' })
+
+    second_request.post!
+    expect(second_request.raw_response.body).to eq('success')
   end
 end